chore: web3 utils folder
